requirements/base.txt
Django==1.11
django-ckeditor==5.4.0
django-extensions==2.0.0
django-filebrowser-no-grappelli==3.7.4
django-geoposition==0.3.0
django-js-asset==1.0.0
django-modeladmin-reorder==0.3
django-uuslug==1.1.8
olefile==0.45.1
Pillow==4.1.1
psycopg2==2.7.4
python-slugify==1.2.4
pytz==2018.3
singlemodeladmin==0.8
six==1.11.0
typing==3.6.4
Unidecode==1.0.22


